requirements.txt
ttkbootstrap
pyinstaller
datetime
pillow
pillow-heif